@@ -93,9 +93,9 @@ class SlideActionBuilderDelegate extends SlideActionDelegate {
   ///
   /// The [builder] must not be null. The [actionCount] argument must not be positive.
   const SlideActionBuilderDelegate({
-    @required this.builder,
-    @required this.actionCount,
-  }) : assert(actionCount != null && actionCount >= 0);
+    required this.builder,
+    required this.actionCount,
+  }) : assert(actionCount >= 0);
 
   /// Called to build slide actions.
   ///
@@ -128,11 +128,11 @@ class SlideActionListDelegate extends SlideActionDelegate {
   ///
   /// The [actions] argument must not be null.
   const SlideActionListDelegate({
-    @required this.actions,
+    required this.actions,
   });
 
   /// The slide actions.
-  final List<Widget> actions;
+  final List<Widget>? actions;
 
   /// The number of actions.
   @override
@@ -141,14 +141,14 @@ class SlideActionListDelegate extends SlideActionDelegate {
   @override
   Widget build(BuildContext context, int index, Animation<double> animation,
           SlidableRenderingMode step) =>
-      actions[index];
+      actions![index];
 }
 
 class _SlidableScope extends InheritedWidget {
   _SlidableScope({
-    Key key,
-    @required this.state,
-    @required Widget child,
+    Key? key,
+    required this.state,
+    required Widget child,
   }) : super(key: key, child: child);
 
   final SlidableState state;
@@ -160,20 +160,20 @@ class _SlidableScope extends InheritedWidget {
 /// The data used by a [Slidable].
 class SlidableData extends InheritedWidget {
   SlidableData({
-    Key key,
-    @required this.actionType,
-    @required this.renderingMode,
-    @required this.totalActionsExtent,
-    @required this.dismissThreshold,
-    @required this.dismissible,
-    @required this.actionDelegate,
-    @required this.overallMoveAnimation,
-    @required this.actionsMoveAnimation,
-    @required this.dismissAnimation,
-    @required this.slidable,
-    @required this.actionExtentRatio,
-    @required this.direction,
-    @required Widget child,
+    Key? key,
+    required this.actionType,
+    required this.renderingMode,
+    required this.totalActionsExtent,
+    required this.dismissThreshold,
+    required this.dismissible,
+    required this.actionDelegate,
+    required this.overallMoveAnimation,
+    required this.actionsMoveAnimation,
+    required this.dismissAnimation,
+    required this.slidable,
+    required this.actionExtentRatio,
+    required this.direction,
+    required Widget child,
   }) : super(key: key, child: child);
 
   /// The type of slide action that is currently been showed by the [Slidable].
@@ -193,16 +193,16 @@ class SlidableData extends InheritedWidget {
   final bool dismissible;
 
   /// The current actions that have to be shown.
-  final SlideActionDelegate actionDelegate;
+  final SlideActionDelegate? actionDelegate;
 
   /// Animation for the whole movement.
   final Animation<double> overallMoveAnimation;
 
   /// Animation for the actions.
-  final Animation<double> actionsMoveAnimation;
+  final Animation<double>? actionsMoveAnimation;
 
   /// Dismiss animation.
-  final Animation<double> dismissAnimation;
+  final Animation<double>? dismissAnimation;
 
   /// The slidable.
   final Slidable slidable;
@@ -233,16 +233,16 @@ class SlidableData extends InheritedWidget {
 
   /// If the [direction] is horizontal, returns the [totalActionsExtent]
   /// otherwise null.
-  double get actionPaneWidthFactor =>
+  double? get actionPaneWidthFactor =>
       directionIsXAxis ? totalActionsExtent : null;
 
   /// If the [direction] is vertical, returns the [totalActionsExtent]
   /// otherwise null.
-  double get actionPaneHeightFactor =>
+  double? get actionPaneHeightFactor =>
       directionIsXAxis ? null : totalActionsExtent;
 
   /// The data from the closest instance of this class that encloses the given context.
-  static SlidableData of(BuildContext context) {
+  static SlidableData? of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<SlidableData>();
   }
 
@@ -258,9 +258,9 @@ class SlidableData extends InheritedWidget {
 
   /// Creates a positioned related to the current direction and showed actions.
   Positioned createPositioned({
-    Widget child,
-    double extent,
-    double position,
+    required Widget child,
+    double? extent,
+    double? position,
   }) {
     return Positioned(
       left: directionIsXAxis ? (showActions ? position : null) : 0.0,
@@ -275,9 +275,9 @@ class SlidableData extends InheritedWidget {
 
   /// Creates a [FractionallyAlignedSizedBox] related to the current direction and showed actions.
   FractionallyAlignedSizedBox createFractionallyAlignedSizedBox({
-    Widget child,
-    double extentFactor,
-    double positionFactor,
+    required Widget child,
+    double? extentFactor,
+    double? positionFactor,
   }) {
     return FractionallyAlignedSizedBox(
       leftFactor:
@@ -297,10 +297,10 @@ class SlidableData extends InheritedWidget {
   List<Widget> buildActions(BuildContext context) {
     return List.generate(
       actionCount,
-      (int index) => actionDelegate.build(
+      (int index) => actionDelegate!.build(
         context,
         index,
-        actionsMoveAnimation,
+        actionsMoveAnimation!,
         SlidableRenderingMode.slide,
       ),
     );
@@ -334,22 +334,22 @@ class SlidableController {
   });
 
   /// Function called when the animation changed.
-  final ValueChanged<Animation<double>> onSlideAnimationChanged;
+  final ValueChanged<Animation<double>?>? onSlideAnimationChanged;
 
   /// Function called when the [Slidable] open status changed.
-  final ValueChanged<bool> onSlideIsOpenChanged;
+  final ValueChanged<bool>? onSlideIsOpenChanged;
 
-  bool _isSlideOpen;
+  bool? _isSlideOpen;
 
-  Animation<double> _slideAnimation;
+  Animation<double>? _slideAnimation;
 
-  SlidableState _activeState;
+  SlidableState? _activeState;
 
   /// The state of the active [Slidable].
-  SlidableState get activeState => _activeState;
+  SlidableState? get activeState => _activeState;
 
   /// Changes the state of the active [Slidable].
-  set activeState(SlidableState value) {
+  set activeState(SlidableState? value) {
     _activeState?._flingAnimationController();
 
     _activeState = value;
@@ -360,19 +360,19 @@ class SlidableController {
         _slideAnimation?.addListener(_handleSlideIsOpenChanged);
         if (_slideAnimation == null) {
           _isSlideOpen = false;
-          onSlideIsOpenChanged(_isSlideOpen);
+          onSlideIsOpenChanged!(_isSlideOpen!);
         }
       }
-      onSlideAnimationChanged(value?.overallMoveAnimation);
+      onSlideAnimationChanged!(value?.overallMoveAnimation);
     }
   }
 
   void _handleSlideIsOpenChanged() {
     if (onSlideIsOpenChanged != null && _slideAnimation != null) {
-      final bool isOpen = _slideAnimation.value != 0.0;
+      final bool isOpen = _slideAnimation!.value != 0.0;
       if (isOpen != _isSlideOpen) {
         _isSlideOpen = isOpen;
-        onSlideIsOpenChanged(_isSlideOpen);
+        onSlideIsOpenChanged!(_isSlideOpen!);
       }
     }
   }
@@ -401,20 +401,20 @@ class Slidable extends StatefulWidget {
   /// state of the dismissed item. Using keys causes the widgets to sync
   /// according to their keys and avoids this pitfall.
   Slidable({
-    Key key,
-    @required Widget child,
-    @required Widget actionPane,
-    List<Widget> actions,
-    List<Widget> secondaryActions,
+    Key? key,
+    required Widget child,
+    required Widget? actionPane,
+    List<Widget>? actions,
+    List<Widget>? secondaryActions,
     double showAllActionsThreshold = 0.5,
     double actionExtentRatio = _kActionsExtentRatio,
     Duration movementDuration = _kMovementDuration,
     Axis direction = Axis.horizontal,
     bool closeOnScroll = true,
     bool enabled = true,
-    SlidableDismissal dismissal,
-    SlidableController controller,
-    double fastThreshold,
+    SlidableDismissal? dismissal,
+    SlidableController? controller,
+    double? fastThreshold,
   }) : this.builder(
           key: key,
           child: child,
@@ -449,9 +449,9 @@ class Slidable extends StatefulWidget {
   /// state of the dismissed item. Using keys causes the widgets to sync
   /// according to their keys and avoids this pitfall.
   Slidable.builder({
-    Key key,
-    @required this.child,
-    @required this.actionPane,
+    Key? key,
+    required this.child,
+    required this.actionPane,
     this.actionDelegate,
     this.secondaryActionDelegate,
     this.showAllActionsThreshold = 0.5,
@@ -462,21 +462,13 @@ class Slidable extends StatefulWidget {
     this.enabled = true,
     this.dismissal,
     this.controller,
-    double fastThreshold,
-  })  : assert(actionPane != null),
-        assert(direction != null),
-        assert(
-            showAllActionsThreshold != null &&
-                showAllActionsThreshold >= .0 &&
+    double? fastThreshold,
+  })  : assert(showAllActionsThreshold >= .0 &&
                 showAllActionsThreshold <= 1.0,
             'showAllActionsThreshold must be between 0.0 and 1.0'),
-        assert(
-            actionExtentRatio != null &&
-                actionExtentRatio >= .0 &&
+        assert(actionExtentRatio >= .0 &&
                 actionExtentRatio <= 1.0,
             'actionExtentRatio must be between 0.0 and 1.0'),
-        assert(closeOnScroll != null),
-        assert(enabled != null),
         assert(dismissal == null || key != null,
             'a key must be provided if slideToDismissDelegate is set'),
         assert(fastThreshold == null || fastThreshold >= .0,
@@ -488,21 +480,21 @@ class Slidable extends StatefulWidget {
   final Widget child;
 
   /// The controller that tracks the active [Slidable] and keep only one open.
-  final SlidableController controller;
+  final SlidableController? controller;
 
   /// A delegate that builds slide actions that appears when the child has been dragged
   /// down or to the right.
-  final SlideActionDelegate actionDelegate;
+  final SlideActionDelegate? actionDelegate;
 
   /// A delegate that builds slide actions that appears when the child has been dragged
   /// up or to the left.
-  final SlideActionDelegate secondaryActionDelegate;
+  final SlideActionDelegate? secondaryActionDelegate;
 
   /// The action pane that controls how the slide actions are animated;
-  final Widget actionPane;
+  final Widget? actionPane;
 
   /// A delegate that controls how to dismiss the item.
-  final SlidableDismissal dismissal;
+  final SlidableDismissal? dismissal;
 
   /// Relative ratio between one slide action and the extent of the child.
   final double actionExtentRatio;
@@ -536,8 +528,8 @@ class Slidable extends StatefulWidget {
   final double fastThreshold;
 
   /// The state from the closest instance of this class that encloses the given context.
-  static SlidableState of(BuildContext context) {
-    final _SlidableScope scope =
+  static SlidableState? of(BuildContext context) {
+    final _SlidableScope? scope =
         context.dependOnInheritedWidgetOfExactType<_SlidableScope>();
     return scope?.state;
   }
@@ -576,23 +568,23 @@ class SlidableState extends State<Slidable>
     )..addStatusListener(_handleShowAllActionsStatusChanged);
   }
 
-  AnimationController _overallMoveController;
+  late AnimationController _overallMoveController;
   Animation<double> get overallMoveAnimation => _overallMoveController.view;
 
-  Animation<double> _actionsMoveAnimation;
-  Animation<double> _dismissAnimation;
+  Animation<double>? _actionsMoveAnimation;
+  Animation<double>? _dismissAnimation;
 
-  AnimationController _resizeController;
-  Animation<double> _resizeAnimation;
+  AnimationController? _resizeController;
+  Animation<double>? _resizeAnimation;
 
   double _dragExtent = 0.0;
 
   SlidableRenderingMode _renderingMode = SlidableRenderingMode.none;
   SlidableRenderingMode get renderingMode => _renderingMode;
 
-  ScrollPosition _scrollPosition;
+  ScrollPosition? _scrollPosition;
   bool _dragUnderway = false;
-  Size _sizePriorToCollapse;
+  Size? _sizePriorToCollapse;
   bool _dismissing = false;
 
   SlideActionType _actionType = SlideActionType.primary;
@@ -610,7 +602,7 @@ class SlidableState extends State<Slidable>
     if (widget.dismissal == null)
       return _kDismissThreshold;
     else
-      return widget.dismissal.dismissThresholds[actionType] ??
+      return widget.dismissal!.dismissThresholds[actionType] ??
           _kDismissThreshold;
   }
 
@@ -619,11 +611,11 @@ class SlidableState extends State<Slidable>
   @override
   bool get wantKeepAlive =>
       !widget.closeOnScroll &&
-      (_overallMoveController?.isAnimating == true ||
+      (_overallMoveController.isAnimating == true ||
           _resizeController?.isAnimating == true);
 
   /// The current actions that have to be shown.
-  SlideActionDelegate get _actionDelegate =>
+  SlideActionDelegate? get _actionDelegate =>
       actionType == SlideActionType.primary
           ? widget.actionDelegate
           : widget.secondaryActionDelegate;
@@ -631,7 +623,7 @@ class SlidableState extends State<Slidable>
   bool get _directionIsXAxis => widget.direction == Axis.horizontal;
 
   double get _overallDragAxisExtent {
-    final Size size = context.size;
+    final Size size = context.size!;
     return _directionIsXAxis ? size.width : size.height;
   }
 
@@ -660,13 +652,13 @@ class SlidableState extends State<Slidable>
     if (widget.closeOnScroll) {
       _scrollPosition = Scrollable.of(context)?.position;
       if (_scrollPosition != null)
-        _scrollPosition.isScrollingNotifier.addListener(_isScrollingListener);
+        _scrollPosition!.isScrollingNotifier.addListener(_isScrollingListener);
     }
   }
 
   void _removeScrollingNotifierListener() {
     if (_scrollPosition != null) {
-      _scrollPosition.isScrollingNotifier.removeListener(_isScrollingListener);
+      _scrollPosition!.isScrollingNotifier.removeListener(_isScrollingListener);
     }
   }
 
@@ -682,7 +674,7 @@ class SlidableState extends State<Slidable>
   /// Opens the [Slidable].
   /// By default it's open the [SlideActionType.primary] action pane, but you
   /// can modify this by setting [actionType].
-  void open({SlideActionType actionType}) {
+  void open({SlideActionType? actionType}) {
     widget.controller?.activeState = this;
 
     if (actionType != null && _actionType != actionType) {
@@ -719,13 +711,13 @@ class SlidableState extends State<Slidable>
   /// Dismisses this [Slidable].
   /// By default it's dismiss by showing the [SlideActionType.primary] action pane, but you
   /// can modify this by setting [actionType].
-  void dismiss({SlideActionType actionType}) {
+  void dismiss({SlideActionType? actionType}) {
     if (_dismissible) {
       _dismissing = true;
       actionType ??= _actionType;
       if (actionType != _actionType) {
         setState(() {
-          this.actionType = actionType;
+          this.actionType = actionType!;
         });
       }
 
@@ -737,7 +729,7 @@ class SlidableState extends State<Slidable>
     if (!widget.closeOnScroll || _scrollPosition == null) return;
 
     // When a scroll starts close this.
-    if (_scrollPosition.isScrollingNotifier.value) {
+    if (_scrollPosition!.isScrollingNotifier.value) {
       close();
     }
   }
@@ -746,25 +738,25 @@ class SlidableState extends State<Slidable>
     _dragUnderway = true;
     widget.controller?.activeState = this;
     _dragExtent =
-        _actionsMoveAnimation.value * _actionsDragAxisExtent * _dragExtent.sign;
+        _actionsMoveAnimation!.value * _actionsDragAxisExtent * _dragExtent.sign;
     if (_overallMoveController.isAnimating) {
       _overallMoveController.stop();
     }
   }
 
   void _handleDragUpdate(DragUpdateDetails details) {
-    if (widget.controller != null && widget.controller.activeState != this) {
+    if (widget.controller != null && widget.controller!.activeState != this) {
       return;
     }
 
-    final double delta = details.primaryDelta;
+    final double delta = details.primaryDelta!;
     _dragExtent += delta;
     setState(() {
       actionType = _dragExtent.sign >= 0
           ? SlideActionType.primary
           : SlideActionType.secondary;
       if (_actionCount > 0) {
-        if (_dismissible && !widget.dismissal.dragDismissible) {
+        if (_dismissible && !widget.dismissal!.dragDismissible) {
           // If the widget is not dismissible by dragging, clamp drag result
           // so the widget doesn't slide past [_totalActionsExtent].
           _overallMoveController.value =
@@ -779,12 +771,12 @@ class SlidableState extends State<Slidable>
   }
 
   void _handleDragEnd(DragEndDetails details) {
-    if (widget.controller != null && widget.controller.activeState != this) {
+    if (widget.controller != null && widget.controller!.activeState != this) {
       return;
     }
 
     _dragUnderway = false;
-    final double velocity = details.primaryVelocity;
+    final double velocity = details.primaryVelocity!;
     final bool shouldOpen = velocity.sign == _dragExtent.sign;
     final bool fast = velocity.abs() > widget.fastThreshold;
 
@@ -801,7 +793,7 @@ class SlidableState extends State<Slidable>
       } else {
         close();
       }
-    } else if (_actionsMoveAnimation.value >= widget.showAllActionsThreshold) {
+    } else if (_actionsMoveAnimation!.value >= widget.showAllActionsThreshold) {
       open();
     } else {
       close();
@@ -837,8 +829,8 @@ class SlidableState extends State<Slidable>
       if (status == AnimationStatus.completed &&
           _overallMoveController.value == _overallMoveController.upperBound &&
           !_dragUnderway) {
-        if (widget.dismissal.onWillDismiss == null ||
-            await widget.dismissal.onWillDismiss(actionType)) {
+        if (widget.dismissal!.onWillDismiss == null ||
+            await widget.dismissal!.onWillDismiss!(actionType)) {
           _startResizeAnimation();
         } else {
           _dismissing = false;
@@ -855,19 +847,17 @@ class SlidableState extends State<Slidable>
 
   void _handleDismiss() {
     widget.controller?.activeState = null;
-    final SlidableDismissal dismissal = widget.dismissal;
+    final SlidableDismissal dismissal = widget.dismissal!;
     if (dismissal.onDismissed != null) {
-      assert(actionType != null);
-      dismissal.onDismissed(actionType);
+      dismissal.onDismissed!(actionType);
     }
   }
 
   void _startResizeAnimation() {
-    assert(_overallMoveController != null);
     assert(_overallMoveController.isCompleted);
     assert(_resizeController == null);
     assert(_sizePriorToCollapse == null);
-    final SlidableDismissal dismissal = widget.dismissal;
+    final SlidableDismissal dismissal = widget.dismissal!;
     if (dismissal.resizeDuration == null) {
       _handleDismiss();
     } else {
@@ -875,22 +865,22 @@ class SlidableState extends State<Slidable>
           AnimationController(duration: dismissal.resizeDuration, vsync: this)
             ..addListener(_handleResizeProgressChanged)
             ..addStatusListener((AnimationStatus status) => updateKeepAlive());
-      _resizeController.forward();
+      _resizeController!.forward();
       setState(() {
         _renderingMode = SlidableRenderingMode.resize;
         _sizePriorToCollapse = context.size;
         _resizeAnimation = Tween<double>(begin: 1.0, end: 0.0).animate(
             CurvedAnimation(
-                parent: _resizeController, curve: _kResizeTimeCurve));
+                parent: _resizeController!, curve: _kResizeTimeCurve));
       });
     }
   }
 
   void _handleResizeProgressChanged() {
-    if (_resizeController.isCompleted) {
+    if (_resizeController!.isCompleted) {
       _handleDismiss();
     } else {
-      widget.dismissal.onResize?.call();
+      widget.dismissal!.onResize?.call();
     }
   }
 
@@ -898,27 +888,27 @@ class SlidableState extends State<Slidable>
   Widget build(BuildContext context) {
     super.build(context); // See AutomaticKeepAliveClientMixin.
 
-    Widget content = widget.child;
+    Widget? content = widget.child;
 
     if (!(!widget.enabled ||
         ((widget.actionDelegate == null ||
-                widget.actionDelegate.actionCount == 0) &&
+                widget.actionDelegate!.actionCount == 0) &&
             (widget.secondaryActionDelegate == null ||
-                widget.secondaryActionDelegate.actionCount == 0)))) {
+                widget.secondaryActionDelegate!.actionCount == 0)))) {
       if (actionType == SlideActionType.primary &&
               widget.actionDelegate != null &&
-              widget.actionDelegate.actionCount > 0 ||
+              widget.actionDelegate!.actionCount > 0 ||
           actionType == SlideActionType.secondary &&
               widget.secondaryActionDelegate != null &&
-              widget.secondaryActionDelegate.actionCount > 0) {
+              widget.secondaryActionDelegate!.actionCount > 0) {
         if (_dismissible) {
           content = widget.dismissal;
 
           if (_resizeAnimation != null) {
             // we've been dragged aside, and are now resizing.
             assert(() {
-              if (_resizeAnimation.status != AnimationStatus.forward) {
-                assert(_resizeAnimation.status == AnimationStatus.completed);
+              if (_resizeAnimation!.status != AnimationStatus.forward) {
+                assert(_resizeAnimation!.status == AnimationStatus.completed);
                 throw FlutterError(
                     'A dismissed Slidable widget is still part of the tree.\n'
                     'Make sure to implement the onDismissed handler and to immediately remove the Slidable\n'
@@ -928,11 +918,11 @@ class SlidableState extends State<Slidable>
             }());
 
             content = SizeTransition(
-              sizeFactor: _resizeAnimation,
+              sizeFactor: _resizeAnimation!,
               axis: _directionIsXAxis ? Axis.vertical : Axis.horizontal,
               child: SizedBox(
-                width: _sizePriorToCollapse.width,
-                height: _sizePriorToCollapse.height,
+                width: _sizePriorToCollapse!.width,
+                height: _sizePriorToCollapse!.height,
                 child: content,
               ),
             );
